+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 swiftmailer extension Change Log
 2.0.6 under development
 -----------------------
 
-- no changes in this release.
+- Enh #27: Added ability to specify message signature (klimov-paul)
 
 
 2.0.5 March 17, 2016